requirements.txt
crawl4ai>=0.8.6
anthropic>=0.40.0
openai>=1.50.0
pyyaml>=6.0
pydantic>=2.0.0
crewai>=1.1.0
fastapi>=0.110.0
uvicorn>=0.27.0
